+++ SignNow.Net/Model/ComplexTags/DateValidatorTag.cs
@@ -1,0 +1,30 @@
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+using SignNow.Net.Internal.Helpers.Converters;
+
+namespace SignNow.Net.Model.ComplexTags
+{
+    public class DateValidatorTag : ComplexTagWithLabel
+    {
+        /// <summary>
+        /// Field type
+        /// </summary>
+        [JsonProperty("type")]
+        [JsonConverter(typeof(StringEnumConverter))]
+        public override FieldType Type { get; protected set; } = FieldType.Text;
+
+        /// <summary>
+        /// Lock Signing Date option
+        /// </summary>
+        [JsonProperty("lsd", Order = 1)]
+        [JsonConverter(typeof(BoolToStringYNJsonConverter))]
+        public bool LockSigningDate { get; set; }
+
+        /// <summary>
+        /// Data validation format for a field
+        /// </summary>
+        [JsonProperty("validator_id", Order = 2)]
+        [JsonConverter(typeof(StringEnumConverter))]
+        public DataValidator Validator { get; set; }
+    }
+}