+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 ## [Unreleased] - TBD
 
 
+## [1.2.1] - 2023-07-13
+### Added
+- Added `DateValidatorTag` to Complex Tags
+- Added `OwnerName` property to SignNowDocument model
+
+### Fixed
+- Fixed Json property name for attachment complex tag
+
+
 ## [1.2.0] - 2023-07-03
 ### Added
 - Added Readme and License files to Nuget package